+++ src/video/webgl/compositors/quad_compositor.js
@@ -1,5 +1,4 @@
 import Vector2d from "../../../math/vector2.js";
-import GLShader from "../glshader.js";
 import { isPowerOfTwo } from "../../../math/math.js";
 import quadVertex from "./../shaders/quad.vert";
 import quadFragment from "./../shaders/quad.frag";
@@ -31,15 +30,15 @@ var V_ARRAY = [
                 {name: "aVertex", size: 2, type: renderer.gl.FLOAT, normalized: false, offset: 0 * Float32Array.BYTES_PER_ELEMENT},
                 {name: "aRegion", size: 2, type: renderer.gl.FLOAT, normalized: false, offset: 2 * Float32Array.BYTES_PER_ELEMENT},
                 {name: "aColor",  size: 4, type: renderer.gl.UNSIGNED_BYTE, normalized: true, offset: 4 * Float32Array.BYTES_PER_ELEMENT}
-            ]
+            ],
+            shader: {
+                vertex: quadVertex, fragment: quadFragment
+            }
         });
 
         // list of active texture units
         this.currentTextureUnit = -1;
         this.boundTextures = [];
-
-        // Load and create shader programs
-        this.quadShader = new GLShader(this.gl, quadVertex, quadFragment);
     }
 
     /**
@@ -57,30 +56,27 @@ var V_ARRAY = [
             }
         }
         this.currentTextureUnit = -1;
-
-        // set the quad shader as the default program
-        this.useShader(this.quadShader);
     }
 
     /**
      * Create a WebGL texture from an image
      * @param {number} unit - Destination texture unit
-     * @param {Image|HTMLCanvasElement|ImageData|Uint8Array[]|Float32Array[]} image - Source image
+     * @param {Image|HTMLCanvasElement|ImageData|Uint8Array[]|Float32Array[]} [pixels=null] - Source image
      * @param {number} filter - gl.LINEAR or gl.NEAREST
      * @param {string} [repeat="no-repeat"] - Image repeat behavior (see {@link ImageLayer#repeat})
-     * @param {number} [w=image.width] - Source image width (Only use with UInt8Array[] or Float32Array[] source image)
-     * @param {number} [h=image.height] - Source image height (Only use with UInt8Array[] or Float32Array[] source image)
-     * @param {number} [b=0] - Source image border width. Must be 0.
+     * @param {number} [w=pixels.width] - Source image width (Only use with UInt8Array[] or Float32Array[] source image)
+     * @param {number} [h=pixels.height] - Source image height (Only use with UInt8Array[] or Float32Array[] source image)
      * @param {boolean} [premultipliedAlpha=true] - Multiplies the alpha channel into the other color channels
      * @param {boolean} [mipmap=true] - Whether mipmap levels should be generated for this texture
      * @returns {WebGLTexture} a WebGL texture
      */
-    createTexture2D(unit, image, filter, repeat = "no-repeat", w = image.width, h = image.height, b = 0, premultipliedAlpha = true, mipmap = true) {
+    createTexture2D(unit, pixels = null, filter, repeat = "no-repeat", w = pixels.width, h = pixels.height, premultipliedAlpha = true, mipmap = true) {
         var gl = this.gl;
         var isPOT = isPowerOfTwo(w) && isPowerOfTwo(h);
-        var texture = gl.createTexture();
         var rs = (repeat.search(/^repeat(-x)?$/) === 0) && (isPOT || this.renderer.WebGLVersion > 1) ? gl.REPEAT : gl.CLAMP_TO_EDGE;
         var rt = (repeat.search(/^repeat(-y)?$/) === 0) && (isPOT || this.renderer.WebGLVersion > 1) ? gl.REPEAT : gl.CLAMP_TO_EDGE;
+
+        var texture = gl.createTexture();
 
         this.bindTexture2D(texture, unit);
 
@@ -88,11 +84,18 @@ var V_ARRAY = [
         gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, rt);
         gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, filter);
         gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, filter);
+
         gl.pixelStorei(gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL, premultipliedAlpha);
-        gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, w, h, b, gl.RGBA, gl.UNSIGNED_BYTE, image);
+
+        if (pixels === null || typeof pixels.byteLength !== "undefined") {
+            // if pixels is undefined, or if it's Uint8Array/Float32Array TypedArray
+            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, w, h, 0, gl.RGBA, gl.UNSIGNED_BYTE, pixels, 0);
+        } else {
+            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, pixels);
+        }
 
         // generate the sprite mimap (used when scaling) if a PowerOfTwo texture
-        if (isPOT && mipmap !== false) {
+        if (isPOT && mipmap === true) {
             gl.generateMipmap(gl.TEXTURE_2D);
         }
 
@@ -165,7 +168,7 @@ var V_ARRAY = [
     /**
      * @ignore
      */
-    uploadTexture(texture, w, h, b, force = false) {
+    uploadTexture(texture, w, h, force = false) {
         var unit = this.renderer.cache.getUnit(texture);
         var texture2D = this.boundTextures[unit];
 
@@ -177,7 +180,6 @@ var V_ARRAY = [
                 texture.repeat,
                 w,
                 h,
-                b,
                 texture.premultipliedAlpha
             );
         } else {
@@ -211,7 +213,7 @@ var V_ARRAY = [
         // upload and activate the texture if necessary
         var unit = this.uploadTexture(texture);
         // set fragement sampler accordingly
-        this.quadShader.setUniform("uSampler", unit);
+        this.defaultShader.setUniform("uSampler", unit);
 
         // Transform vertices
         var m = this.viewMatrix,